biginteger

+++ base/access.py
@@ -1,11 +1,11 @@
 from .postgress import metadata
 
-from sqlalchemy import Table, Column, Integer, String 
+from sqlalchemy import Table, Column, Integer, String, BigInteger
 
 t_access = Table(
     "access",
     metadata,
-    Column('u_id', Integer), # номер юзера
+    Column('u_id', BigInteger), # номер юзера
     Column('c_id', Integer), # номер команды
     Column('comment', String), # комментарий
     )
