+++ base/choice.py
@@ -1,11 +1,11 @@
 from .postgress import metadata
 
-from sqlalchemy import Table, Column, Integer
+from sqlalchemy import Table, Column, Integer, BigInteger
 
 t_choice = Table(
     "choice",
     metadata,
-    Column('u_id', Integer),
+    Column('u_id', BigInteger),
     Column('c_id', Integer),
         )
